fsmonitor: Use fsmonitor data in `git diff`
This makes use of the fsmonitor extension to skip lstat() calls on
files that fsmonitor judged as unmodified.

The fsmonitor test is modified to make but one call to `git add`,
because `git add` uses the `git diff` machinery, and the "hook" that
is installed in testing reports a fixed set of files as "changed" upon
each invocation.  With three calls to `git add`, only the last file
added would report as being fresh, as its fsmonitor invocation would
have reported the earlier-added ones as now "dirty."