fsmonitor: Copy fsmonitor data when creating indexes in `git checkout`
`git checkout HEAD` creates a new, empty, index and and fills it from
the data in the commit at HEAD.  In creating a new index, it did not
copy the last update time from the fsmonitor extension, though the
index entries themselves were copied, preserving their
CE_FSMONITOR_VALID bit.  However, the writing of the fsmonitor
extension itself is predicated on a non-zero last update time; thus,
the new index was written without the extension, and the next `git
status` had to resort to lstat()'ing every file.

Copy the fsmonitor_last_update field over onto the new index.  Index
entries for files on disk which were changed are marked as 'dirty';
though we know that their index entries are "valid", the timestamp
that will be written out in the extension predates the checkout, and
thus the next fsmonitor query will return those files -- and thus they
will be simply marked "dirty" again.  For this reason, we do not
bother marking _all_ entries as valid.